chore(TS) export filter types (#8913)
