+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [next]
 
 - BREAKING fix(): reflect NUM_FRACTION_DIGITS to toObject path data [#8909](https://github.com/fabricjs/fabric.js/pull/8909)
+- chore(lint) export filter colors and brushes types [#8913](https://github.com/fabricjs/fabric.js/pull/8913)
 - chore(lint) Add a rule for import type [#8907](https://github.com/fabricjs/fabric.js/pull/8907)
 - fix(Object): dirty unflagging inconsistency [#8910](https://github.com/fabricjs/fabric.js/pull/8910)
 - chore(TS): minor type/import fixes [#8904](https://github.com/fabricjs/fabric.js/pull/8904)
