+++ src/brushes/SprayBrush.ts
@@ -5,13 +5,7 @@ import { Rect } from '../shapes/Rect';
 import { getRandomInt } from '../util/internals';
 import type { Canvas } from '../canvas/Canvas';
 import { BaseBrush } from './BaseBrush';
-
-export type SprayBrushPoint = {
-  x: number;
-  y: number;
-  width: number;
-  opacity: number;
-};
+import type { SprayBrushPoint } from './typedefs';
 
 /**
  *
