+++ src/filters/index.ts
@@ -3,3 +3,6 @@ export * as filters from './filters';
 export { getFilterBackend, initFilterBackend } from './FilterBackend';
 export { Canvas2dFilterBackend } from './Canvas2dFilterBackend';
 export { WebGLFilterBackend } from './WebGLFilterBackend';
+export { isWebGLPipelineState } from './utils';
+
+export * from './typedefs';
